# libs/contract/openapi.yaml
openapi: 3.0.3
info:
  title: Ori HTTP Contract
  description: REST API contract for Ori runtime, replacing the legacy JSON-RPC surface.
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Default development server
paths:
  /health:
    get:
      summary: Health probe
      operationId: getHealth
      responses:
        '200':
          description: Backend is ready to accept requests
          content:
            text/plain:
              schema:
                type: string
                example: ok
        default:
          $ref: '#/components/responses/ErrorResponse'
  /resources:
    get:
      summary: List saved resource definitions
      operationId: listResources
      responses:
        '200':
          description: Collection of saved resources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcesResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /resources/connect:
    post:
      summary: Start (or refresh) a resource session for a resource
      operationId: connectResource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceConnectRequest'
      responses:
        '201':
          description: Connection attempt accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceConnectResult'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorPayload'
        '409':
          description: Connection is already in progress and cannot be restarted yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorPayload'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /resources/{resourceName}/nodes:
    get:
      summary: Fetch schema nodes for a resource
      operationId: getNodes
      parameters:
        - name: resourceName
          in: path
          required: true
          description: Name of the resource to inspect
          schema:
            type: string
        - name: nodeId
          in: query
          required: false
          description: Optional repeated parameter limiting the nodes returned
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        '200':
          description: Matching schema nodes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodesResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorPayload'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /queries:
    post:
      summary: Execute a SQL query asynchronously
      operationId: execQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryExecRequest'
      responses:
        '202':
          description: Query accepted and job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryExecResponse'
        '404':
          description: Resource not found or no active resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorPayload'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /queries/{jobId}/cancel:
    post:
      summary: Cancel a running query job
      operationId: cancelQuery
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Cancel request accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required:
                  - status
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorPayload'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /queries/{jobId}/result:
    get:
      summary: Retrieve a previously stored query result view
      operationId: getQueryResult
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Result view for the job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResultResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorPayload'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /events:
    get:
      summary: Subscribe to server-sent events for resource and query notifications
      operationId: streamEvents
      responses:
        '200':
          description: Stream of events in SSE format
          content:
            text/event-stream:
              schema:
                type: string
        default:
          $ref: '#/components/responses/ErrorResponse'
components:
  responses:
    ErrorResponse:
      description: Generic error payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorPayload'
  schemas:
    PasswordConfig:
      type: object
      properties:
        type:
          type: string
          description: Password provider type
          enum: [plain_text, shell, keychain]
        key:
          type: string
          description: Provider-specific identifier (plain text value, shell command, or keychain account)
      required:
        - type
        - key
    TlsConfig:
      type: object
      properties:
        mode:
          type: string
          nullable: true
          description: TLS mode (e.g. require, verify-ca, verify-full)
        caCertPath:
          type: string
          nullable: true
          description: CA certificate path
        certPath:
          type: string
          nullable: true
          description: Client certificate path
        keyPath:
          type: string
          nullable: true
          description: Client key path
    Resource:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        host:
          type: string
          nullable: true
        port:
          type: integer
          nullable: true
        database:
          type: string
        username:
          type: string
          nullable: true
        password:
          $ref: '#/components/schemas/PasswordConfig'
        tls:
          $ref: '#/components/schemas/TlsConfig'
      required:
        - name
        - type
        - database
    ResourcesResponse:
      type: object
      properties:
        resources:
          type: array
          items:
            $ref: '#/components/schemas/Resource'
      required:
        - resources
    ResourceConnectRequest:
      type: object
      properties:
        resourceName:
          type: string
      required:
        - resourceName
    ResourceConnectResult:
      type: object
      properties:
        result:
          type: string
          enum: [success, fail, connecting]
        userMessage:
          type: string
          nullable: true
      required:
        - result
    NodeEdge:
      type: object
      properties:
        items:
          type: array
          items:
            type: string
        truncated:
          type: boolean
      required:
        - items
        - truncated
    DatabaseNodeAttributes:
      type: object
      additionalProperties: false
      properties:
        resource:
          type: string
        engine:
          type: string
        isDefault:
          type: boolean
        file:
          type: string
        sequence:
          type: integer
        pageSize:
          type: integer
          format: int64
        encoding:
          type: string
      required:
        - resource
        - engine
        - isDefault
    SchemaNodeAttributes:
      type: object
      additionalProperties: false
      properties:
        resource:
          type: string
        engine:
          type: string
        isDefault:
          type: boolean
      required:
        - resource
        - engine
        - isDefault
    TableNodeAttributes:
      type: object
      additionalProperties: false
      properties:
        resource:
          type: string
        table:
          type: string
        tableType:
          type: string
        definition:
          type: string
      required:
        - resource
        - table
        - tableType
    ViewNodeAttributes:
      type: object
      additionalProperties: false
      properties:
        resource:
          type: string
        table:
          type: string
        tableType:
          type: string
        definition:
          type: string
      required:
        - resource
        - table
        - tableType
    ColumnNodeAttributes:
      type: object
      additionalProperties: false
      properties:
        resource:
          type: string
        table:
          type: string
        column:
          type: string
        ordinal:
          type: integer
        dataType:
          type: string
        notNull:
          type: boolean
        defaultValue:
          type: string
        primaryKeyPosition:
          type: integer
        charMaxLength:
          type: integer
          format: int64
        numericPrecision:
          type: integer
          format: int64
        numericScale:
          type: integer
          format: int64
      required:
        - resource
        - table
        - column
        - ordinal
        - dataType
        - notNull
    ConstraintNodeAttributes:
      type: object
      additionalProperties: false
      properties:
        resource:
          type: string
        table:
          type: string
        constraintName:
          type: string
        constraintType:
          type: string
        columns:
          type: array
          items:
            type: string
        referencedTable:
          type: string
        referencedDatabase:
          type: string
        referencedSchema:
          type: string
        referencedColumns:
          type: array
          items:
            type: string
        onUpdate:
          type: string
        onDelete:
          type: string
        match:
          type: string
        indexName:
          type: string
        checkClause:
          type: string
      required:
        - resource
        - table
        - constraintName
        - constraintType
    IndexNodeAttributes:
      type: object
      additionalProperties: false
      properties:
        resource:
          type: string
        table:
          type: string
        indexName:
          type: string
        unique:
          type: boolean
        primary:
          type: boolean
        columns:
          type: array
          items:
            type: string
        includeColumns:
          type: array
          items:
            type: string
        definition:
          type: string
        method:
          type: string
        predicate:
          type: string
      required:
        - resource
        - table
        - indexName
        - unique
        - primary
    TriggerNodeAttributes:
      type: object
      additionalProperties: false
      properties:
        resource:
          type: string
        table:
          type: string
        triggerName:
          type: string
        timing:
          type: string
        orientation:
          type: string
        events:
          type: array
          items:
            type: string
        statement:
          type: string
        condition:
          type: string
        enabledState:
          type: string
        definition:
          type: string
      required:
        - resource
        - table
        - triggerName
        - timing
        - orientation
    NodeBase:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        edges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/NodeEdge'
      required:
        - id
        - name
        - edges
    DatabaseNode:
      allOf:
        - $ref: '#/components/schemas/NodeBase'
        - type: object
          properties:
            type:
              type: string
              enum: [database]
            attributes:
              $ref: '#/components/schemas/DatabaseNodeAttributes'
          required:
            - type
            - attributes
    SchemaNode:
      allOf:
        - $ref: '#/components/schemas/NodeBase'
        - type: object
          properties:
            type:
              type: string
              enum: [schema]
            attributes:
              $ref: '#/components/schemas/SchemaNodeAttributes'
          required:
            - type
            - attributes
    TableNode:
      allOf:
        - $ref: '#/components/schemas/NodeBase'
        - type: object
          properties:
            type:
              type: string
              enum: [table]
            attributes:
              $ref: '#/components/schemas/TableNodeAttributes'
          required:
            - type
            - attributes
    ViewNode:
      allOf:
        - $ref: '#/components/schemas/NodeBase'
        - type: object
          properties:
            type:
              type: string
              enum: [view]
            attributes:
              $ref: '#/components/schemas/ViewNodeAttributes'
          required:
            - type
            - attributes
    ColumnNode:
      allOf:
        - $ref: '#/components/schemas/NodeBase'
        - type: object
          properties:
            type:
              type: string
              enum: [column]
            attributes:
              $ref: '#/components/schemas/ColumnNodeAttributes'
          required:
            - type
            - attributes
    ConstraintNode:
      allOf:
        - $ref: '#/components/schemas/NodeBase'
        - type: object
          properties:
            type:
              type: string
              enum: [constraint]
            attributes:
              $ref: '#/components/schemas/ConstraintNodeAttributes'
          required:
            - type
            - attributes
    IndexNode:
      allOf:
        - $ref: '#/components/schemas/NodeBase'
        - type: object
          properties:
            type:
              type: string
              enum: [index]
            attributes:
              $ref: '#/components/schemas/IndexNodeAttributes'
          required:
            - type
            - attributes
    TriggerNode:
      allOf:
        - $ref: '#/components/schemas/NodeBase'
        - type: object
          properties:
            type:
              type: string
              enum: [trigger]
            attributes:
              $ref: '#/components/schemas/TriggerNodeAttributes'
          required:
            - type
            - attributes
    Node:
      oneOf:
        - $ref: '#/components/schemas/DatabaseNode'
        - $ref: '#/components/schemas/SchemaNode'
        - $ref: '#/components/schemas/TableNode'
        - $ref: '#/components/schemas/ViewNode'
        - $ref: '#/components/schemas/ColumnNode'
        - $ref: '#/components/schemas/ConstraintNode'
        - $ref: '#/components/schemas/IndexNode'
        - $ref: '#/components/schemas/TriggerNode'
      discriminator:
        propertyName: type
        mapping:
          database: '#/components/schemas/DatabaseNode'
          schema: '#/components/schemas/SchemaNode'
          table: '#/components/schemas/TableNode'
          view: '#/components/schemas/ViewNode'
          column: '#/components/schemas/ColumnNode'
          constraint: '#/components/schemas/ConstraintNode'
          index: '#/components/schemas/IndexNode'
          trigger: '#/components/schemas/TriggerNode'
    NodesResponse:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/Node'
      required:
        - nodes
    QueryExecOptions:
      type: object
      properties:
        maxRows:
          type: integer
          minimum: 1
      additionalProperties: false
    QueryExecRequest:
      type: object
      properties:
        resourceName:
          type: string
        jobId:
          type: string
          format: uuid
        query:
          type: string
        params:
          oneOf:
            - type: object
              additionalProperties: {}
            - type: array
              items: {}
        options:
          $ref: '#/components/schemas/QueryExecOptions'
      required:
        - resourceName
        - jobId
        - query
    QueryExecResponse:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [running, failed]
        message:
          type: string
          nullable: true
      required:
        - jobId
        - status
    QueryResultColumn:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
      required:
        - name
        - type
    QueryResultResponse:
      type: object
      properties:
        columns:
          type: array
          items:
            $ref: '#/components/schemas/QueryResultColumn'
        rows:
          type: array
          items:
            type: array
            items: {}
        rowCount:
          type: integer
        truncated:
          type: boolean
        rowsAffected:
          type: integer
          nullable: true
      required:
        - columns
        - rows
        - rowCount
        - truncated
    ErrorPayload:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: {}
      required:
        - code
        - message